requirements.txt
nltk==3.5
numpy==1.18.3
lxml==4.5.2
pytorch_lightning==1.1.1
PyYAML==5.3.1
scikit_learn==0.23.2
torch==1.7.1
torchnlp==0.0.0.1
torchvision==0.8.2
transformers==4.0.1
tensorflow
